requirements.txt
flask
yt-dlp
openai
python-dotenv
gunicorn